note core relocation
