+++ README.md
@@ -131,6 +131,7 @@ curl -X POST http://localhost:8000/report
 - Added a new simplified React frontend under `frontend` that uses the
   `uui-theme-eduverse_dark` styles and shows report data in a table.
 - Added a Dockerfile and `nginx.conf` so the frontend can run via Docker Compose.
+- Moved `core` package outside `app` for simpler imports.
 
 ## 👤 Maintainer
